fix printf string format

+++ tonemapper_unittest.cpp
@@ -30,8 +30,8 @@ int main(int argc, char *argv[]) {
     auto t2 = std::chrono::high_resolution_clock::now();
     auto img = tonemapper.Map(imgIn, false);
     auto t3 = std::chrono::high_resolution_clock::now();
-    printf("%lld ms\n", std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1).count());
-    printf("%lld ms\n", std::chrono::duration_cast<std::chrono::milliseconds>(t3 - t2).count());
+    printf("%lld ms\n", (long long)std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1).count());
+    printf("%lld ms\n", (long long)std::chrono::duration_cast<std::chrono::milliseconds>(t3 - t2).count());
 
     quink::ImageStore::StoreImage("tonemap.png", quink::ImageFormat::PngImage, std::move(img));
